+++ landbosse/excelio/XlsxGenerator.py
@@ -1,13 +1,10 @@
 import xlsxwriter
 from xlsxwriter.utility import xl_rowcol_to_cell
 import pandas as pd
-import numpy as np
 import os
 import traceback
 
-from .filename_functions import timestamp_filename
-from .filename_functions import landbosse_output_dir
-from .filename_functions import landbosse_input_dir
+from .XlsxFileOperations import XlsxFileOperations
 
 
 class XlsxGenerator:
@@ -22,7 +19,7 @@ class XlsxGenerator:
     output Excel workbook.
     """
 
-    def __init__(self, output_xlsx):
+    def __init__(self, output_xlsx, file_ops):
         """
         This constructor sets the name of the .xlsx file for writing
 
@@ -35,6 +32,9 @@ class XlsxGenerator:
             The name of the .xlsx file to write. Do not include the .xlsx at
             the end of the filename. Also, this filename will be timestamped
             before it is written.
+
+        file_ops : XlsxFileOperations
+            An instance of XlsxFileOperations to manage file names.
         """
 
         # Set all instance attributes to None first in the constructor as good
@@ -43,7 +43,8 @@ class XlsxGenerator:
         self.header_format = None
         self.scientific_format = None
         self.percent_format = None
-        self.output_xlsx_path = timestamp_filename(landbosse_output_dir(), output_xlsx, 'xlsx')
+        self.output_xlsx_path = os.path.join(file_ops.landbosse_output_dir(), f'{output_xlsx}.xlsx')
+        self.file_ops = file_ops
 
     def __enter__(self):
         """
@@ -149,7 +150,7 @@ class XlsxGenerator:
         This writes a detailed outputs tab. It takes a list of dictionaries
         as the parameters and in each of those dictionaries it looks at the keys:
 
-        ['project_id', 'module', 'type', 'variable_df_key_col_name', 'unit', 'value']
+        ['project_id', 'module', 'type', 'variable_df_key_col_name', 'unit', 'numeric value', 'non_numeric_value']
 
         The values of each of those keys become each cell in the row.
 
@@ -163,117 +164,53 @@ class XlsxGenerator:
         worksheet.set_column(4, 4, 17)
         worksheet.set_column(5, 5, 66)
         worksheet.set_column(0, 2, 17)
-        for idx, col_name in enumerate(['project_id', 'module', 'type', 'variable_df_key_col_name', 'unit', 'value', 'last number']):
+
+        for idx, col_name in enumerate(['Project ID', 'Module', 'Variable or DataFrame', 'name', 'unit', 'Numeric value', 'Non-numeric value']):
             worksheet.write(0, idx, col_name, self.header_format)
+
+        # Go through each row and create Excel rows from each of those rows.
         for row_idx, row in enumerate(rows):
             worksheet.write(row_idx + 1, 0, row['project'])
             worksheet.write(row_idx + 1, 1, row['module'])
             worksheet.write(row_idx + 1, 2, row['type'])
             worksheet.write(row_idx + 1, 3, row['variable_df_key_col_name'])
             worksheet.write(row_idx + 1, 4, row['unit'])
-            if type(row['value']) is str or type(row['value']) is int or type(row['value']) is float:
-                worksheet.write(row_idx + 1, 5, row['value'], self.scientific_format)
+
+            value = row['value']
+            value_is_number = self._is_numeric(value)
+            if value_is_number:
+                worksheet.write(row_idx + 1, 5, value, self.scientific_format)
             else:
-                worksheet.write(row_idx + 1, 5, str(row['value']))
+                worksheet.write(row_idx + 1, 6, value)
+
+            # If there is a last_number, which means this is a dataframe row that has a number
+            # at the end, write this into the numeric value column.
+
             if 'last_number' in row:
-                if type(row['last_number']) is int or type(row['last_number']) is float:
-                    worksheet.write(row_idx + 1, 6, row['last_number'], self.scientific_format)
-                else:
-                    worksheet.write(row_idx + 1, 6, str(row['last_number']))
+                worksheet.write(row_idx + 1, 5, row['last_number'], self.scientific_format)
+
         worksheet.freeze_panes(1, 0)  # Freeze the first row.
 
-    def tab_details_with_validation(self, rows, validation_xlsx):
+    def _is_numeric(self, value):
         """
-        This is a special method that is for validation only. If you
-        just need the details tab, use the tab_details() method.
+        This method tests if a value is a numeric (that is, can be parsed
+        by float()) or non numeric (which cannot be parsed).
 
-        This writes an .xlsx file. It takes a list of dictionaries as the parameter
-        and in each of those dictionaries it looks at the keys:
-
-        ['project_id', 'module', 'type', 'variable_df_key_col_name', 'unit', 'value']
-
-        as the order of columns for each row of the .xlsx file.
-
-        It also sets up some number and text formatting and column widths for
-        the .xlsx file.
-
-        It timestamps the output filename of the output xlsx.
+        The decision from this method determines whether values go into
+        the numeric or non-numeric columns.
 
         Parameters
         ----------
-        rows : list
-            list of dicts. See above.
-
-        validation_xlsx : str
-            The name of .xlsx file which contains the validation data. It is
-            assumed to live in the inputs directory. Please include .xlsx
+        value
+            The value to be tested.
 
         Returns
         -------
-        str
-            The string of the full pathname to the file just written.
+        bool
+            True if the value is numeric, False otherwise.
         """
-        # Read the validation inputs. Validation data assumed to be on 'Sheet1'
-        validation_xlsx_path = os.path.join(landbosse_input_dir(), validation_xlsx)
-        validation = pd.read_excel(validation_xlsx_path, 'Sheet1')
-        validation_row_dicts = [row.to_dict() for _, row in validation.iterrows()]
-
-        # Make a worksheet for the model outputs
-        worksheet = self.workbook.add_worksheet('details')
-
-        # Write values to the output spreadsheet.
-        for idx, col_name in enumerate(['Project ID', 'module', 'type', 'variable_df_key_col_name', 'unit', 'value', 'last_number', 'expected_value', 'difference', 'primary_key']):
-            worksheet.write(0, idx, col_name, self.header_format)
-        row_idx = 1
-        for row in rows:
-            pk_part_1_cell = xl_rowcol_to_cell(row_idx, 0)
-            pk_part_2_cell = xl_rowcol_to_cell(row_idx, 3)
-            value_cell = xl_rowcol_to_cell(row_idx, 6) if 'last_number' in row else xl_rowcol_to_cell(row_idx, 5)
-            expected_value_cell = xl_rowcol_to_cell(row_idx, 7)
-            worksheet.write(row_idx, 0, row['project'])
-            worksheet.write(row_idx, 1, row['module'])
-            worksheet.write(row_idx, 2, row['type'])
-            worksheet.write(row_idx, 3, row['variable_df_key_col_name'])
-            worksheet.write(row_idx, 4, row['unit'])
-
-            if type(row['value']) is int or type(row['value']) is float:
-                worksheet.write(row_idx, 5, row['value'], self.scientific_format)
-            else:
-                worksheet.write(row_idx, 5, str(row['value']))
-            if 'last_number' in row:
-                if type(row['last_number']) is int or type(row['last_number']) is float:
-                    worksheet.write(row_idx, 6, row['last_number'], self.scientific_format)
-                else:
-                    worksheet.write(row_idx, 6, str(row['last_number']))
-
-            f = '=INDEX(details_validation!$C$2:$C${}, MATCH(details!J{}, details_validation!$J$2:$J${}, 0))'.format(len(validation_row_dicts) + 1, row_idx + 1, len(validation_row_dicts) + 1)
-            worksheet.write(row_idx, 7, f, self.scientific_format)
-
-            worksheet.write(row_idx, 8, '={}/{}'.format(expected_value_cell, value_cell), self.percent_format)
-            worksheet.write_formula(row_idx, 9, '=CONCATENATE({}, {})'.format(pk_part_1_cell, pk_part_2_cell))
-            row_idx += 1
-        worksheet.freeze_panes(1, 0)  # Freeze the first row.
-
-        # Write the validation outputs to a tab in the excel file. Make the primary key
-        worksheet = self.workbook.add_worksheet('details_validation')
-        validations_col_names = list(validation_row_dicts[0].keys())[:]
-        validations_col_names.append('primary_key')
-        for idx, col_name in enumerate(validations_col_names):
-            worksheet.write(0, idx, col_name, self.header_format)
-            worksheet.set_column(idx, idx, 50)
-        row_idx = 1
-        for row in validation_row_dicts:
-            pk_part_1_cell = xl_rowcol_to_cell(row_idx, 0)
-            pk_part_2_cell = xl_rowcol_to_cell(row_idx, 5)
-            worksheet.write(row_idx, 0, row['Project ID'])
-            worksheet.write(row_idx, 1, row['variable_name_validation_data'])
-            worksheet.write(row_idx, 2, row['expected_value'])
-            worksheet.write(row_idx, 3, row['module'])
-            worksheet.write(row_idx, 4, row['type'])
-            worksheet.write(row_idx, 5, row['variable_df_key_col_name'])
-            worksheet.write(row_idx, 6, row['unit'])
-            worksheet.write(row_idx, 7, row['value'])
-            worksheet.write(row_idx, 8, row['notes'])
-            worksheet.write(row_idx, 9, '=CONCATENATE({}, {})'.format(pk_part_1_cell, pk_part_2_cell))
-            row_idx += 1
-        worksheet.freeze_panes(1, 0)  # Freeze the first row.
+        try:
+            float(value)
+        except ValueError:
+            return False
+        return True